Removed some unneeded files(mostly duplicate queries).  Also updated queries to get the user token from the state.  Moved some of the queries we created to correct spots in the folder directory.  The ony main thing that could break with this is the queries not getting the right token again.
