+++ src/node_modules_prev_reaapp/store/selectors/auth.tsx
@@ -1,4 +1,4 @@
-/**
+/**s
  * Selector logic for accessing auth data in view components
  *
  * @package: REA app
