+++ src/queries/Agent/AgentOrganizations.ts
@@ -9,8 +9,6 @@
 // import { connect } from 'react-redux'
 // import { gql, graphql, compose } from 'react-apollo'
 
-// import { AppState } from '@vflows/store/types'
-// import { getActiveLoginToken } from '@vflows/store/selectors/auth'
 
 // import { coreAgentFields, coreOrganizationFields } from '../_fragments/Agent'
 
@@ -31,9 +29,9 @@
 
 // export default compose(
 //   // bind input data from the store
-//   connect((state: AppState) => ({
+//   connect((state) => ({
 //     variables: {
-//       token: getActiveLoginToken(state),
+//       token: state.getUserInfo.currentUserToken,
 //     },
 //   })),
 //   graphql(query, {
