+++ src/queries/EconomicEvent/getEconomicEventById.tsx
@@ -8,9 +8,6 @@
 import { connect } from "react-redux";
 import { graphql, compose } from "react-apollo";
 import gql from "graphql-tag";
-import { AppState } from "@vflows/store/types";
-import { getActiveLoginToken } from "@vflows/store/selectors/auth";
-import { create } from "domain";
 import economicEventInterface from "./EconomicEvent.tsx";
 
 /**
@@ -52,7 +49,7 @@ export default compose(
    */
   connect(state => ({
     variables: {
-      token: getActiveLoginToken(state)
+      token: state.getUserInfo.currentUserToken
     }
   })),
