+++ src/queries/Process/process.tsx
@@ -9,8 +9,6 @@
 import { connect } from 'react-redux'
 import { graphql, compose } from "react-apollo";
 import gql from "graphql-tag";
-import { AppState } from '@vflows/store/types'
-import { getActiveLoginToken } from '@vflows/store/selectors/auth'
 
 import { coreCommitmentFields, coreEventFields } from '../_fragments/Process'
 
@@ -107,9 +105,9 @@ export interface AgentType {
 
 export default compose(
   // bind input data from the store
-  connect((state: AppState) => ({
+  connect((state) => ({
     variables: {
-      token: getActiveLoginToken(state),
+      token: state.getUserInfo.currentUserToken,
     },
   })),
   graphql(query, {
