+++ src/queries/ProcessClassification/getAllProcessClassifications.tsx
@@ -10,8 +10,6 @@ import { connect } from "react-redux";
 import { graphql, compose } from "react-apollo";
 import gql from "graphql-tag";
 
-import { AppState } from "@vflows/store/types.js";
-import { getActiveLoginToken } from "@vflows/store/selectors/auth.js";
 
 import { processClassificationInterface } from "./processClassificationInterface";
 
@@ -31,9 +29,9 @@ ${processClassificationInterface}
 
 export default compose(
   // Bind input data from the store
-  connect((state: AppState) => ({
+  connect((state) => ({
     variables: {
-      token: getActiveLoginToken(state),
+      token: state.getUserInfo.currentUserToken,
     },
   })),
