+++ src/queries/ProcessClassification/getProcessClassificationById.tsx
@@ -9,8 +9,6 @@
 import { connect } from "react-redux";
 import { graphql, compose } from "react-apollo";
 import gql from "graphql-tag";
-import { AppState } from "@vflows/store/types.js";
-import { getActiveLoginToken } from "@vflows/store/selectors/auth.js";
 
 import { processClassificationInterface } from "./processClassificationInterface";
 
@@ -30,9 +28,9 @@ ${processClassificationInterface}
 
 export default compose(
   // Bind input data from the store
-  connect((state: AppState) => ({
+  connect((state) => ({
     variables: {
-      token: getActiveLoginToken(state),
+      token: state.getUserInfo.currentUserToken,
     },
   })),
