+++ src/queries/User/UserEmailExists.tsx
@@ -6,10 +6,8 @@
  */
 
 
-import {connect} from "react-redux";
 import { graphql, compose } from "react-apollo";
 import gql from "graphql-tag";
-import {AppState} from "@vflows/store/types.js";
 
 
 export const query = gql`
