+++ src/routes.js
@@ -3,7 +3,6 @@ import {Switch, Route} from 'react-router-dom';
 import OrganizationCatalog from './pages/OrganizationCatalog.js';
 import BasicTextPage from './pages/BasicTextPage.js';
 import ErrorPage from './pages/ErrorPage.js';
-import OrganizationRegistration from './pages/OrganizationRegistration.js';
 import LoginPage from './pages/LoginPage.js';
 import SearchResults from './pages/oldReaapp/organizationCatalog'
 
@@ -15,7 +14,6 @@ class Routes extends Component {
                     <Route exact path="/" component={OrganizationCatalog}/>
                     <Route exact path="/login" component={LoginPage}/>
                     <Route exact path="/text" component={BasicTextPage}/>
-                    <Route path="/registration/organization" component={OrganizationRegistration}/>
                     <Route exact path = "/test" component={SearchResults}/>
                     <Route component={ErrorPage}/>
                 </Switch>
